+++ aegisub/audio_provider_hd.h
@@ -58,5 +58,5 @@ public:
 	HDAudioProvider(AudioProvider *source);
 	~HDAudioProvider();
 
-	void GetAudio(void *buf, __int64 start, __int64 count);
+	void GetAudio(void *buf, long long start, long long count);
 };
